requirements.txt
SQLAlchemy
pandas
tensorflow
notebook
pyodbc
mysql-connector-python
python-dotenv
matplotlib